move scipy_sparse to arrays
